Modularize updater.rs into submodules for maintainability
Split the 581-line updater.rs into focused submodules:
- check.rs: Update checking logic via GitHub Releases API
- apply.rs: Update application and binary replacement
- version.rs: Binary version detection for daemon auto-restart
- staged.rs: Staged update system for safe binary replacement
- background.rs: Background update checking and download
- tests.rs: Unit tests

All new files are under 200 lines, well within the 500-line limit.
